// src/redux/action-types.ts
export const enum Actions{
    CHECK_PASSWORDS = "CHECK_PASSWORDS",
    LOG_USER_API = "LOG_USER_API",
    CLEAR_USER = "CLEAR_USER",
    GET_CURRENT_USER = "GET_CURRENT_USER",
    CLEAR_CURRENT_USER = "CLEAR_CURRENT_USER",
    VALID_PASSWORD = "VALID_PASSWORD",
    GET_USER = "GET_USER",
    REG_USER = "REG_USER",
    SET_LOADING = "SET_LOADING",
    LOG_USER = "LOG_USER",
    LOG_CURRENT_USER = "LOG_CURRENT_USER",
    SET_ALLOW_LOGIN = "SET_ALLOW_LOGIN",
    DROP_ALLOW_LOGIN = "DROP_ALLOW_LOGIN",
    CHECK_LOGIN = "CHECK_LOGIN",
    SET_ALLOW_NICKNAME = "SET_ALLOW_NICKNAME",
    CHECK_NICKNAME = "CHECK_NICKNAME",
    DROP_ALLOW_NICKNAME = "DROP_ALLOW_NICKNAME",
    SET_ALLOW_PASSWORD = "SET_ALLOW_PASSWORD",
    DROP_ALLOW_PASSWORD = "DROP_ALLOW_PASSWORD",
    DROP_ALL = "DROP_ALL"
}